+++ tests/unit/difference/test_policyshard.py
@@ -156,14 +156,14 @@ DIFFERENCE_SCENARIOS = {
     "proper_subset_with_condition": {
         "first": PolicyShard(
             effect="Allow",
-            effective_action=EffectiveAction(inclusion=Action("s3:*")),
+            effective_action=EffectiveAction(inclusion=Action("*")),
             effective_resource=EffectiveResource(inclusion=Resource("*")),
             effective_principal=EffectivePrincipal(Principal("AWS", "*")),
             conditions=frozenset(),
         ),
         "second": PolicyShard(
-            effect="Allow",
-            effective_action=EffectiveAction(inclusion=Action("s3:*")),
+            effect="Deny",
+            effective_action=EffectiveAction(inclusion=Action("*")),
             effective_resource=EffectiveResource(inclusion=Resource("*")),
             effective_principal=EffectivePrincipal(Principal("AWS", "arn:aws:iam::123456789012:root")),
             conditions=frozenset({Condition("Key", "Operator", ["Value"])}),
@@ -171,34 +171,38 @@ DIFFERENCE_SCENARIOS = {
         "result": [
             PolicyShard(
                 effect="Allow",
-                effective_action=EffectiveAction(inclusion=Action("s3:*")),
-                effective_resource=EffectiveResource(inclusion=Resource("*")),
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
                 effective_principal=EffectivePrincipal(
-                    Principal("AWS", "*"), frozenset({Principal("AWS", "arn:aws:iam::123456789012:root")})
+                    inclusion=Principal(type="AWS", value="*"),
+                    exclusions=frozenset({Principal(type="AWS", value="arn:aws:iam::123456789012:root")}),
                 ),
                 conditions=frozenset(),
+                not_conditions=frozenset(),
             ),
             PolicyShard(
                 effect="Allow",
-                effective_action=EffectiveAction(inclusion=Action("s3:*")),
-                effective_resource=EffectiveResource(inclusion=Resource("*")),
-                effective_principal=EffectivePrincipal(Principal("AWS", "arn:aws:iam::123456789012:root")),
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
+                ),
                 conditions=frozenset(),
-                not_conditions=frozenset({Condition("Key", "Operator", ["Value"])}),
+                not_conditions=frozenset({Condition(key="Key", operator="Operator", values=["Value"])}),
             ),
         ],
     },
-    "exact_match_on_inclusion_with_exclusion": {
+    "exact_match_on_inclusion_with_exclusion_and_condition": {
         "first": PolicyShard(
             effect="Allow",
-            effective_action=EffectiveAction(inclusion=Action("s3:*")),
+            effective_action=EffectiveAction(inclusion=Action("*")),
             effective_resource=EffectiveResource(inclusion=Resource("*")),
             effective_principal=EffectivePrincipal(Principal("AWS", "*")),
             conditions=frozenset(),
         ),
         "second": PolicyShard(
             effect="Allow",
-            effective_action=EffectiveAction(inclusion=Action("s3:*")),
+            effective_action=EffectiveAction(inclusion=Action("*")),
             effective_resource=EffectiveResource(
                 inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::DOC-EXAMPLE-BUCKET/*")})
             ),
@@ -208,20 +212,95 @@ DIFFERENCE_SCENARIOS = {
         "result": [
             PolicyShard(
                 effect="Allow",
-                effective_action=EffectiveAction(inclusion=Action("s3:*")),
-                effective_resource=EffectiveResource(inclusion=Resource("arn:aws:s3:::DOC-EXAMPLE-BUCKET/*")),
-                effective_principal=EffectivePrincipal(Principal("AWS", "*")),
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(
+                    inclusion=Resource("arn:aws:s3:::DOC-EXAMPLE-BUCKET/*"), exclusions=frozenset()
+                ),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
+                ),
                 conditions=frozenset(),
+                not_conditions=frozenset(),
             ),
             PolicyShard(
                 effect="Allow",
-                effective_action=EffectiveAction(inclusion=Action("s3:*")),
-                effective_resource=EffectiveResource(
-                    inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::DOC-EXAMPLE-BUCKET/*")})
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
                 ),
-                effective_principal=EffectivePrincipal(Principal("AWS", "*")),
                 conditions=frozenset(),
-                not_conditions=frozenset({Condition("Key", "Operator", ["Value"])}),
+                not_conditions=frozenset({Condition(key="Key", operator="Operator", values=["Value"])}),
+            ),
+        ],
+    },
+    "simple_deny_with_exclusion": {
+        "first": PolicyShard(
+            effect="Allow",
+            effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+            effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
+            effective_principal=EffectivePrincipal(inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()),
+        ),
+        "second": PolicyShard(
+            effect="Deny",
+            effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+            effective_resource=EffectiveResource(
+                inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
+            ),
+            effective_principal=EffectivePrincipal(inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()),
+            not_conditions=frozenset(),
+        ),
+        "result": [
+            PolicyShard(
+                effect="Allow",
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(
+                    inclusion=Resource("arn:aws:s3:::examplebucket/*"), exclusions=frozenset()
+                ),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
+                ),
+                conditions=frozenset(),
+                not_conditions=frozenset(),
+            )
+        ],
+    },
+    "deny_action_and_resource_subsets": {
+        "first": PolicyShard(
+            effect="Allow",
+            effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+            effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
+            effective_principal=EffectivePrincipal(inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()),
+        ),
+        "second": PolicyShard(
+            effect="Deny",
+            effective_action=EffectiveAction(inclusion=Action("s3:PutObject"), exclusions=frozenset()),
+            effective_resource=EffectiveResource(inclusion=Resource("arn:aws:s3:::examplebucket/*")),
+            effective_principal=EffectivePrincipal(inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()),
+            not_conditions=frozenset(),
+        ),
+        "result": [
+            PolicyShard(
+                effect="Allow",
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset({Action("s3:PutObject")})),
+                effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
+                ),
+                conditions=frozenset(),
+                not_conditions=frozenset(),
+            ),
+            PolicyShard(
+                effect="Allow",
+                effective_action=EffectiveAction(inclusion=Action("*"), exclusions=frozenset()),
+                effective_resource=EffectiveResource(
+                    inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
+                ),
+                effective_principal=EffectivePrincipal(
+                    inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
+                ),
+                conditions=frozenset(),
+                not_conditions=frozenset(),
             ),
         ],
     },